requirements.txt
python-telegram-bot==21.4
httpx==0.27.0
asyncio
aiosqlite
